chore(docs): move image list to fixture library
