+++ apps/insight-viewer-docs/containers/Measurement/Viewer/index.tsx
@@ -9,8 +9,7 @@ import InsightViewer, {
   MeasurementMode,
   Measurement,
 } from '@lunit/insight-viewer'
-import { IMAGES } from '../../../const'
-import { RULER_MEASUREMENTS, CIRCLE_MEASUREMENTS } from '@insight-viewer-library/fixtures'
+import { IMAGES, RULER_MEASUREMENTS, CIRCLE_MEASUREMENTS } from '@insight-viewer-library/fixtures'
 
 export type InitialMeasurements = {
   [mode in MeasurementMode]: Measurement[]
