+++ apps/insight-viewer-docs/containers/Viewport/Image1.tsx
@@ -1,10 +1,10 @@
 import { useEffect, useCallback } from 'react'
 import { Box, Stack, Switch, Button } from '@chakra-ui/react'
 import InsightViewer, { useImage, useViewport, Viewport } from '@lunit/insight-viewer'
+import { IMAGES } from '@insight-viewer-library/fixtures'
 import { ViewerWrapper } from '../../components/Wrapper'
 import CustomProgress from '../../components/CustomProgress'
 import OverlayLayer from '../../components/OverlayLayer'
-import { IMAGES } from '../../const'
 import { INITIAL_VIEWPORT1 } from './const'
 
 export default function Image1(): JSX.Element {
