+++ apps/insight-viewer-docs/containers/Viewport/Image2.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useCallback } from 'react'
 import { Box, Stack, Button } from '@chakra-ui/react'
 import InsightViewer, { useImage, useViewport, Viewport } from '@lunit/insight-viewer'
+import { IMAGES } from '@insight-viewer-library/fixtures'
 import { ViewerWrapper } from '../../components/Wrapper'
 import CustomProgress from '../../components/CustomProgress'
 import OverlayLayer from '../../components/OverlayLayer'
 import { INITIAL_VIEWPORT2 } from './const'
-import { IMAGES } from '../../const'
 
 export default function Image1(): JSX.Element {
   const { image } = useImage({
