+++ libs/fixtures/src/annotation/circles.ts
