+++ libs/fixtures/src/annotation/contours.ts
