+++ libs/fixtures/src/annotation/polygons.ts
