+++ libs/fixtures/src/annotation/texts.ts
