+++ libs/fixtures/src/image/index.ts
@@ -1,0 +1,17 @@
+export const IMAGES = [
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000000.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000001.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000002.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000003.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000004.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000005.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000006.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000007.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000008.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000009.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000010.dcm',
+  'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT000011.dcm',
+  'wadouri:https://static.lunit.io/insight/samples/cxr/Nodule.dcm',
+  'wadouri:https://static.lunit.io/insight/samples/mmg/case01_LCC.dcm',
+]
+export const WRONG_IMAGE = 'wadouri:https://static.lunit.io/fixtures/dcm-files/series/CT0000100.dcm'
