+++ libs/fixtures/src/index.ts
@@ -1,7 +1,8 @@
-export * from './lib/circles'
-export * from './lib/contours'
-export * from './lib/freeLines'
-export * from './lib/lines'
-export * from './lib/polygons'
-export * from './lib/rulers'
-export * from './lib/texts'
+export * from './image'
+export * from './annotation/circles'
+export * from './annotation/contours'
+export * from './annotation/freeLines'
+export * from './annotation/lines'
+export * from './annotation/polygons'
+export * from './annotation/rulers'
+export * from './annotation/texts'
